With SwiftPM 5.4 done, rename version-specific tests to make it clear what they're testing and to make room for vNext tests.  Add a "_0" suffix to the tests for major versions so that they sort by increasing tools version.
