+++ Tests/PackageLoadingTests/PDNextVersionLoadingTests.swift
@@ -1,7 +1,7 @@
 /*
  This source file is part of the Swift.org open source project
 
- Copyright (c) 2020 Apple Inc. and the Swift project authors
+ Copyright (c) 2021 Apple Inc. and the Swift project authors
  Licensed under Apache License v2.0 with Runtime Library Exception
 
  See http://swift.org/LICENSE.txt for license information
@@ -21,22 +21,4 @@ class PackageDescriptionNextVersionLoadingTests: PackageDescriptionLoadingTests 
         .vNext
     }
 
-    func testExecutableTargets() throws {
-        let stream = BufferedOutputByteStream()
-        stream <<< """
-            import PackageDescription
-            let package = Package(
-               name: "Foo",
-               targets: [
-                   .executableTarget(
-                       name: "Foo"
-                    ),
-               ]
-            )
-            """
-
-        loadManifest(stream.bytes) { manifest in
-            XCTAssertEqual(manifest.targets[0].type, .executable)
-        }
-    }
 }
